:sparkles: Initial working prototype.
Implements a basic Audio Graph that contains any number of audio nodes as well as tracks how they are connected. Basic oscillator and gain nodes are currently implemented, as well as a function to allow for custom nodes to be created. Additionally, graph to JSON encoding is supported and this extends to audio nodes (including custom nodes).